dns_brute/requirements.txt
requests==2.27.1
dnspython==2.2.1